+++ main1061.cpp
@@ -28,7 +28,7 @@ int main() {
   x5o6F94e  =  (short)1236322931;
   p0QZ_zHoL  =  (short)-986007249;
 
-sKl3j74zH = IWt54_Vt(C0lvc, -781596244);
+sKl3j74zH = IWt54_Vt(C0lvc, (unsigned short)-781596244);
 
 j1WT5kx = (short int)(n4u6DP + (N2DUn7VU_1));
 {
