+++ main1267.cpp
@@ -31,7 +31,7 @@ int main(void) {
     t0R06kM8    =    (short)-1150921103;
     X249nhQ    =    (short)-1470053096;
     OzrY    =    (short)1456959732;
-    cIFib    =    -1567095839;
+    cIFib    =    (short)-1567095839;
     e6J3o9WTJ    =    -850697648;
     V5u77GKg95    =    (short)1836729713;
     Wq1YRFJhL    =    (short)-1505174265;
@@ -41,7 +41,7 @@ int main(void) {
     D8328Za_    =    (short)413589348;
     FH455    =    (short)270529619;
     A56KCgp3rD    =    (short)-1646367054;
-    e9ZsyHiY4W    =    1916146447;
+    e9ZsyHiY4W    =    (short)1916146447;
     hXv8    =    568497499;
     coXP31Sz6    =    (short)181586394;
     LuVR6F    =    (short)-2071510024;
@@ -51,7 +51,7 @@ int main(void) {
     DGpkk    =    (short)-259519719;
     q4_A41    =    (short)-404033332;
     WwJ2e    =    (short)888516614;
-    v27T9H    =    -1386488423;
+    v27T9H    =    (short)-1386488423;
     b42t    =    626356405;
     PsCWv    =    (short)649489189;
     tNx_KMb    =    (short)1396929782;
@@ -61,7 +61,7 @@ int main(void) {
     sXO12p    =    (short)1550098417;
     pclEb    =    (short)-75063585;
     s4QgP24z1O    =    (short)-1989538713;
-    ltcG    =    -79972495;
+    ltcG    =    (short)-79972495;
     awi6899    =    398753268;
     Km5BwyU_r    =    (short)-200118006;
     cc7NouW    =    (short)-163841323;
@@ -71,7 +71,7 @@ int main(void) {
     A_hz    =    (short)-1291331437;
     rFy8h    =    (short)-887000587;
     zI0PD39v    =    (short)-461347975;
-    v2_85A925    =    -1973403086;
+    v2_85A925    =    (short)-1973403086;
     k3UU7E8A    =    1206534121;
     Q5et    =    (short)1700740633;
     B6UeNTPUG    =    (short)-724479105;
@@ -81,7 +81,7 @@ int main(void) {
     r8Y9XdcG8    =    (short)1572086288;
     ERlM1jUj4G    =    (short)1694983658;
     r7_5    =    (short)1575961788;
-    F7W1    =    -787394164;
+    F7W1    =    (short)-787394164;
     qQvm35rf    =    -1711857924;
     VEUQ    =    (short)-762576676;
     H_49wdCF6w    =    (short)1832472978;
